# docs/swagger.yaml
basePath: /api/v1
definitions:
  comment.Comment:
    properties:
      created_at:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - message
    type: object
  comment.CommentCreateRequest:
    properties:
      message:
        type: string
      photo_id:
        type: integer
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  photo.Photo:
    properties:
      caption:
        type: string
      comments:
        items:
          $ref: '#/definitions/comment.Comment'
        type: array
      created_at:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - photo_url
    - title
    type: object
  photo.PhotoCreateRequest:
    properties:
      caption:
        type: string
      photo_url:
        type: string
      title:
        type: string
    type: object
...
  reponse.WebResponseSuccess:
    properties:
      data: {}
      message:
        type: string
    type: object
...
host: localhost:9999
info:
  contact: {}
paths:
  /comment/createcomment:
    post:
      consumes:
      - application/json
      description: how to Create Comment
      parameters:
      - description: Input Data Comment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/comment.CommentCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/comment.Comment'
              type: object
      security:
      - Bearer: []
      summary: create Comment example
      tags:
      - Comments
  /comment/deletecomment/{id}:
    delete:
      consumes:
      - application/json
      description: how to Delete Comment
      parameters:
      - description: id Comment
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reponse.WebResponseSuccess'
      security:
      - Bearer: []
      summary: delete Comment example
      tags:
      - Comments
  /comment/getall:
    get:
      consumes:
      - application/json
      description: how to get all Comment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/comment.Comment'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get All Comment example
      tags:
      - Comments
  /comment/getone/{id}:
    get:
      consumes:
      - application/json
      description: how to get Comment
      parameters:
      - description: id photo
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/comment.Comment'
              type: object
      security:
      - Bearer: []
      summary: Get Comment example
      tags:
      - Comments
  /comment/updatecomment/{id}:
    put:
      consumes:
      - application/json
      description: how to update Comment
      parameters:
      - description: Id Comment
        in: path
        name: id
        type: integer
      - description: Input Data Comment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/comment.CommentCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/comment.Comment'
              type: object
      security:
      - Bearer: []
      summary: update Comment example
      tags:
      - Comments
  /photo/createphoto:
    post:
      consumes:
      - application/json
      description: how to Create photo
      parameters:
      - description: Input Data Photo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/photo.PhotoCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/photo.Photo'
              type: object
      security:
      - Bearer: []
      summary: create Photo example
      tags:
      - Photos
  /photo/deletephoto/{id}:
    delete:
      consumes:
      - application/json
      description: how to Delete photo
      parameters:
      - description: id Photo
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reponse.WebResponseSuccess'
      security:
      - Bearer: []
      summary: delete Photo example
      tags:
      - Photos
  /photo/getall:
    get:
      consumes:
      - application/json
      description: how to get all Photo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/photo.Photo'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get All Photo example
      tags:
      - Photos
  /photo/getone/{id}:
    get:
      consumes:
      - application/json
      description: how to get Photo
      parameters:
      - description: id photo
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/photo.Photo'
              type: object
      security:
      - Bearer: []
      summary: Get Photo example
      tags:
      - Photos
  /photo/updatephoto/{id}:
    put:
      consumes:
      - application/json
      description: how to update photo
      parameters:
      - description: id photo
        in: path
        name: id
        type: integer
      - description: Input Data Photo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/photo.PhotoCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/reponse.WebResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/photo.Photo'
              type: object
      security:
      - Bearer: []
      summary: update Photo example
      tags:
      - Photos
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"